chore: rename src/sass-compiler to src/compiler
