+++ src/index.ts
@@ -2,7 +2,7 @@
 
 export { sassBuild, sassBuildFiles, sassBuildString } from './sass-build';
 export { sassCompile, sassCompileString } from './sass-compile';
-export { wrapImplementation as wrapCompiler } from './sass-compiler';
+export { wrapImplementation as wrapCompiler } from './compiler';
 export { sassCacheImporter } from './importers/cache-importer';
 export { sassPackageImporter } from './importers/package-importer';
 export { cli } from './cli';
